// src/types/DPlayer.d.ts
export interface APIBackend {
    /**
     * @description read danmaku from API backend
     * @param options API backend read options
     */
    read(options: APIBackendReadOptions): void;

    /**
     * @description send danmaku to API backend
     * @param options API backend send options
     */
    send(options: APIBackendSendOptions): void;
}
// ...
export interface APIBackendReadOptions {
    url?: string;
    success: (danmaku: Dan[]) => void;
    error: (message?: string) => void;
}

export interface APIBackendSendOptions {
    url?: string;
    data: Dan;
    success: () => void;
    error: (message?: string) => void;
}

export interface DanmakuItem {
    text: string;
    color: string;
    type: DanmakuType;
    size: DanmakuSize;
    border?: boolean;
}

export interface Dan {
    token?: string,
    id?: string,
    author?: string,
    time: number,
    text: string;
    color: string;
    type: DanmakuType;
    size: DanmakuSize;
}